+++ R/dplyr.R
@@ -168,7 +168,6 @@ tbl_pyspark_sdf <- function(x) {
 
 tbl_temp_name <- function() glue("{temp_prefix()}{random_string()}")
 
-#' @importFrom sparklyr sdf_register
 #' @export
 sdf_register.spark_pyobj <- function(x, name = NULL) {
   sc <- spark_connection(x)
@@ -191,8 +190,41 @@ python_sdf <- function(x) {
 }
 
 python_obj_get <- function(x) {
+  UseMethod("python_obj_get")
+}
+
+python_obj_get.ml_connect_model <- function(x) {
+  x$pipeline$pyspark_obj
+}
+
+python_obj_get.default <- function(x) {
+  if (inherits(x, "character")) {
+    return(x)
+  }
   sc <- spark_connection(x)
-  sc$session
+  if (inherits(sc$session, "python.builtin.object")) {
+    return(sc$session)
+  }
+}
+
+python_obj_get.python.builtin.object <- function(x) {
+  x
+}
+
+python_obj_get.spark_pyobj <- function(x) {
+  x[["pyspark_obj"]]
+}
+
+python_obj_get.ml_connect_model <- function(x) {
+  x[["pipeline"]][["pyspark_obj"]]
+}
+
+python_obj_get.ml_connect_estimator <- function(x) {
+  x[[".jobj"]]
+}
+
+python_obj_get.ml_connect_pipeline_model <- function(x) {
+  x[[".jobj"]]
 }
 
 python_obj_con_set <- function(sc, obj) {
@@ -212,8 +244,30 @@ tbl_pyspark_temp <- function(x, conn, tmp_name = NULL) {
   if (is.null(tmp_name)) {
     tmp_name <- tbl_temp_name()
   }
-  x$createOrReplaceTempView(tmp_name)
+  py_x <- python_obj_get(x)
+  py_x$createOrReplaceTempView(tmp_name)
   tbl(sc, tmp_name)
 }
 
 setOldClass(c("tbl_pyspark", "tbl_spark"))
+
+#' @export
+`[.tbl_pyspark` <- function(x, i) {
+  # this is defined to match the interface to sparklyr::`[.tbl_spark`. But it really
+  # should be more flexible, taking row specs, multiple args, etc. matching
+  # semantics of R dataframes and take advantage of
+  # reticulate::`[.python.builtin.object` for constructing slices, etc.
+  if (is.null(i)) {
+    # special case, since pyspark has no "emptyDataFrame" method to invoke
+    sc <- spark_connection(x)
+
+    pyspark.sql.types <- reticulate::import("pyspark.sql.types")
+    ss <- x$src$state$spark_context # SparkSession obj
+    edf <- ss$createDataFrame(list(), pyspark.sql.types$StructType(list()))
+
+    tmp_name <- tbl_temp_name()
+    edf$createOrReplaceTempView(tmp_name)
+    return(tbl(sc, tmp_name))
+  }
+  NextMethod()
+}