+++ R/python-import-check.R
@@ -3,7 +3,7 @@ import_check <- function(x, envname) {
   env_loaded <- NA
   look_for_env <- TRUE
 
-  if(file.exists(envname)) {
+  if (file.exists(envname)) {
     env_is_file <- TRUE
     env_path <- envname
   } else {
@@ -50,12 +50,13 @@ import_check <- function(x, envname) {
     if (env_found) {
       if (env_loaded) {
         # found & loaded
-        cli_abort(c(
-          paste(
-            "{.emph '{x}' }{.header is not available in the }",
-            "{.emph '{envname}' }{.header Python environment.}"
-          )
-        ), call = NULL)
+        rlang::abort(
+          paste0(
+            "Issue loading Python library '", x, "'\n",
+            as.character(out)
+          ),
+          call = NULL
+        )
       } else {
         cli_abort(c(
           "{.emph '{x}' }{.header is not available current Python environment.}",
@@ -81,10 +82,11 @@ import_check <- function(x, envname) {
       if (look_for_env) {
         msg <- paste(
           "{.header Using the }{.emph '{envname}' }{.header Python}",
-          "{.header environment }{.class ({py_exe()})}"
+          "{.header environment }"
         )
         cli_div(theme = cli_colors())
         cli_alert_success(msg)
+        cli_bullets(c(" " = "{.class Path: {py_exe()}}"))
         cli_end()
       }
     } else {