+++ R/python-install.R
@@ -14,6 +14,10 @@
 #' @param ... Passed on to [`reticulate::py_install()`]
 #' @param as_job Runs the installation if using this function within the
 #' RStudio IDE.
+#' @param install_ml Installs ML related Python libraries. Defaults to TRUE. This
+#' is mainly for machines with limited storage to avoid installing the rather
+#' large 'torch' library if the ML features are not going to be used. This will
+#' apply to any environment backed by 'Spark' version 3.5 or above.
 #' @returns It returns no value to the R session. This function purpose is to
 #' create the 'Python' environment, and install the appropriate set of 'Python'
 #' libraries inside the new environment. During runtime, this function will send
@@ -29,6 +33,7 @@ install_pyspark <- function(
     new_env = TRUE,
     method = c("auto", "virtualenv", "conda"),
     as_job = TRUE,
+    install_ml = FALSE,
     ...) {
   install_as_job(
     libs = "pyspark",
@@ -38,6 +43,7 @@ install_pyspark <- function(
     new_env = new_env,
     method = method,
     as_job = as_job,
+    install_ml = install_ml,
     ... = ...
   )
 }
@@ -57,6 +63,7 @@ install_databricks <- function(
     new_env = TRUE,
     method = c("auto", "virtualenv", "conda"),
     as_job = TRUE,
+    install_ml = FALSE,
     ...) {
   if (!is.null(version) && !is.null(cluster_id)) {
     cli_div(theme = cli_colors())
@@ -69,9 +76,13 @@ install_databricks <- function(
     cli_end()
   }
 
-  if(is.null(envname)) {
+  if (is.null(envname)) {
     if (is.null(version) && !is.null(cluster_id)) {
-      version <- cluster_dbr_version(cluster_id)
+      version <- databricks_dbr_version(
+        cluster_id = cluster_id,
+        host = databricks_host(),
+        token = databricks_token()
+      )
     }
   }
 
@@ -83,6 +94,7 @@ install_databricks <- function(
     new_env = new_env,
     method = method,
     as_job = as_job,
+    install_ml = install_ml,
     ... = ...
   )
 }
@@ -95,6 +107,7 @@ install_as_job <- function(
     new_env = NULL,
     method = c("auto", "virtualenv", "conda"),
     as_job = TRUE,
+    install_ml = TRUE,
     ...) {
   args <- c(as.list(environment()), list(...))
   in_rstudio <- FALSE
@@ -121,6 +134,7 @@ install_as_job <- function(
       python_version = python_version,
       new_env = new_env,
       method = method,
+      install_ml = install_ml,
       ... = ...
     )
   }
@@ -153,6 +167,7 @@ install_environment <- function(
     python_version = NULL,
     new_env = NULL,
     method = c("auto", "virtualenv", "conda"),
+    install_ml = FALSE,
     ...) {
   if (is.null(version)) {
     cli_div(theme = cli_colors())
@@ -188,12 +203,20 @@ install_environment <- function(
         add_torch <- FALSE
       }
       ln <- "databricks"
-      envname <- use_envname(method = "databricks_connect", version = ver_name)
+      envname <- use_envname(
+        method = "databricks_connect",
+        version = ver_name,
+        ask_if_not_installed = FALSE
+      )
     } else {
       if (compareVersion(as.character(ver_name), "3.5") < 0) {
         add_torch <- FALSE
       }
-      envname <- use_envname(method = "spark_connect", version = ver_name)
+      envname <- use_envname(
+        method = "spark_connect",
+        version = ver_name,
+        ask_if_not_installed = FALSE
+      )
     }
     cli_alert_success(
       "Automatically naming the environment:{.emph '{envname}'}"
@@ -208,8 +231,8 @@ install_environment <- function(
     "grpcio_status"
   )
 
-  if (add_torch) {
-    packages <- c(packages, "torch", "torcheval")
+  if (add_torch && install_ml) {
+    packages <- c(packages, pysparklyr_env$ml_libraries)
   }
 
   method <- match.arg(method)