+++ R/utils.R
@@ -6,13 +6,57 @@
 #' @rdname pipe
 #' @keywords internal
 #' @export
-#' @importFrom magrittr %>%
 #' @usage lhs \%>\% rhs
 #' @param lhs A value or the magrittr placeholder.
 #' @param rhs A function call using the magrittr semantics.
 #' @return The result of calling `rhs(lhs)`.
 NULL
 
+reticulate_python_check <- function(ignore = FALSE, unset = TRUE, message = TRUE) {
+  if (ignore) {
+    return("")
+  }
+
+  out <- ""
+
+  env_var <- Sys.getenv("RETICULATE_PYTHON", unset = NA)
+
+  if (current_product_connect() && !is.na(env_var)) {
+    out <- env_var
+  }
+
+  if (!is.na(env_var) && out == "") {
+    if (unset) {
+      Sys.unsetenv("RETICULATE_PYTHON")
+      if (message) {
+        cli_div(theme = cli_colors())
+        cli_alert_warning(paste0(
+          "{.header Your }{.emph 'RETICULATE_PYTHON'}",
+          " {.header environment variable was unset.}"
+        ))
+        cli_bullets(c(
+          " " = "{.class To recover, restart R after you closing your Spark session}"
+        ))
+        cli_end()
+      }
+    } else {
+      if (message) {
+        cli_alert_warning(paste(
+          paste0(
+            "Your {.emph 'RETICULATE_PYTHON'} environment is set, ",
+            "which may cause connectivity issues"
+          ),
+          paste0(
+            "Use {.code Sys.unsetenv(\"RETICULATE_PYTHON\")} to clear for the",
+            "duration of your current R session."
+          ),
+          sep = "\n"
+        ))
+      }
+    }
+  }
+  out
+}
 
 check_arg_supported <- function(x, msg = NULL) {
   arg <- enquo(x)
@@ -44,4 +88,12 @@ cli_internal_abort <- function(msg) {
   cli_div(theme = cli_colors())
   cli_abort(msg, call = NULL)
   cli_end()
+}
+
+current_product_connect <- function() {
+  out <- FALSE
+  if (Sys.getenv("RSTUDIO_PRODUCT") == "CONNECT") {
+    out <- TRUE
+  }
+  out
 }