+++ NEWS.md
@@ -1,4 +1,61 @@
-# pysparklyr (Development)
+# pysparklyr 0.1.2
+
+### Improvements
+
+* When connecting,`spark_connect()`, it will automatically prompt the
+user to install a Python Environment a viable one is not  not found. 
+This way, the R user will not have to run `install_databricks()`/
+`install_pyspark()` manually when using the package for the first time. (#69)
+
+* Instead of simply warning that `RETICULATE_PYTHON` is set, it will now un-set
+the variable. This allows `pysparklyr` to select the correct Python environment.
+It will output a console message to the user when the variable is un-set. (#65).
+Because of how Posit Connect manages `reticulate` Python environments, `pysparklyr`
+will force the use of the Python environment under that particular published
+content's `RETICULATE_PYTHON`.
+
+* Adds enhanced RStudio Snippet for Databricks connections. It will automatically 
+check the cluster's version by pooling the Databricks REST API with the cluster's
+ID. This to check if there is a pre-installed Python environment that will
+suport the cluster's version. All these generate notifications in the snippet's
+UI. It also adds integration with Posit Workbench's new 'Databricks' pane. The
+snippet looks for a specific environment variable that Posit Workbench temporarily
+sets with the value of the cluster ID, and initializes the snippet with that
+value. (#53)
+
+* Adds `install_ml` argument to `install_databricks()` and `install_pyspark()`. 
+The ML related Python libraries are very large, and take a long time to install.
+In most cases, the user will not need these to interact with the cluster. The 
+`install_ml` argument is a flag that will control if the ML libraries will
+be installed. It defaults to `FALSE`. The first time the R user runs an ML 
+related function, then `pysparklyr` will prompt them to install the needed
+libraries at that time.(#63, #78)
+
+* Adds support for Databricks OAuth by adding a handler to the Posit Connect 
+integration. Internally, it centralizes the authentication processing into
+one un-exported function. (#68)
+
+* General improvements to all of console outputs
+
+### Machine Learning
+
+* Adds support for: 
+  - `ft_standard_scaler()`
+  - `ft_max_abs_scaler()`
+  - `ml_logistic_regression()`
+  - `ml_pipeline()`
+  - `ml_save()`
+  - `ml_predict()`
+  - `ml_transform()`
+
+* Adds `ml_prepare_dataset()` in lieu of a Vector Assembler transformer
+
+### Fixes
+
+* Fixes error in use_envname() - No environment name provided, and no 
+environment was automatically identified (#71)
+
+# pysparklyr 0.1.1
 
 ### Improvements